added periodstartdate file
So the application will get the current period from the periodStartDate
text file. Basically it's for the program to know what period it is
currently in. And there should be a user story that allows the user to
change the current period to the next one e.g. from 2014-01-01 the
periodStartDate will become 2014-01-16. We'll implement that for the
next iteration.